# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для Telegram Mini App по разделению счетов",
        "title": "BillSplitter API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
        "/auth/login/telegram": {
            "post": {
                "description": "Создаёт или получает пользователя по Telegram ID и возвращает sessionID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация через Telegram",
                "parameters": [
                    {
                        "description": "Данные пользователя из Telegram",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.LoginTelegramReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.LoginTelegramRes"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, но в debug моде возвращает детали ошибки",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Возвращает данные пользователя, извлечённые по sessionID из контекста",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Получить данные текущего пользователя",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_auth.MeRes"
                        }
                    },
                    "400": {
                        "description": "Пользователь не найден или сессия невалидна",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bills": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bill"
                ],
                "summary": "Получить список чеков ивента",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID ивента",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_transport_http_bill.Bill"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Чек создается от имени пользователя из сессии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bill"
                ],
                "summary": "Создает чек внутри ивента",
                "parameters": [
                    {
                        "description": "Данные чека",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_bill.CreateBillRq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ResponseID"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/billsplitter-monolith_internal_utils_http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "billsplitter-monolith_internal_domain_user.ExtraInfo": {
            "type": "object",
            "properties": {
                "telegram_id": {
                    "type": "integer"
                }
            }
        },
        "billsplitter-monolith_internal_domain_user.User": {
            "type": "object",
            "properties": {
                "extra": {
                    "$ref": "#/definitions/billsplitter-monolith_internal_domain_user.ExtraInfo"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "billsplitter-monolith_internal_utils_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                }
            }
        },
        "billsplitter-monolith_internal_utils_http.ResponseID": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                }
            }
        },
...
        "internal_transport_http_auth.LoginTelegramReq": {
            "type": "object",
            "required": [
                "telegramID"
            ],
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "telegramID": {
                    "type": "integer",
                    "example": 123456789
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
        "internal_transport_http_auth.LoginTelegramRes": {
            "type": "object",
            "properties": {
                "sessionID": {
                    "type": "string",
                    "example": "b42b0a8e-0d1f-4c3d-939f-85fbbdc9be62"
                },
                "userInfo": {
                    "$ref": "#/definitions/billsplitter-monolith_internal_domain_user.User"
                }
            }
        },
        "internal_transport_http_auth.MeRes": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/billsplitter-monolith_internal_domain_user.User"
                }
            }
        },
        "internal_transport_http_bill.Bill": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_user_id": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_bill.Participant"
                    }
                },
                "split_type": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_bill.CreateBillRq": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_bill.ParticipantRq"
                    }
                },
                "split_type": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "integer"
                }
            }
        },
        "internal_transport_http_bill.Participant": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "integer"
                }
            }
        },
        "internal_transport_http_bill.ParticipantRq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "member_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Сессионный токен. Значение из /auth/login/telegram передавайте через Authorize -\u003e X-Session-ID header.",
            "type": "apiKey",
            "name": "X-Session-ID",
            "in": "header"
        }
    }
}